src/requirements.txt
requests
lxml
python-telegram-bot
python-json-logger
coloredlogs